# api/openapi.yaml
openapi: 3.1.0
info:
  title: Orange Cat Investments (OCI) Platform API
  description: API specification for OCI Backend-For-Frontend (BFF) endpoints serving Customer Portal (`customer.oci.local`) and Employee Portal (`employee.oci.local`).
  version: 1.0.0
servers:
  - url: https://api.oci.local/api/v1
    description: Common Ingress Development API Router
  - url: https://customer.oci.local/api/v1
    description: Customer BFF (ZITADEL OIDC Auth)
  - url: https://employee.oci.local/api/v1
    description: Employee BFF (Forgejo OAuth2 Auth)
paths:
  /facilities/assets:
    get:
      summary: List Hardware Assets
      operationId: listAssets
      tags:
        - Facilities Assets
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, maintenance_required, offline, pending_return, decommissioned]
        - name: asset_type
          in: query
          required: false
          schema:
            type: string
            enum: [edge_camera, observation_perch, smart_collar, gateway, feeder]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of hardware assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/HardwareAsset'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500InternalError'

    post:
      summary: Register New Hardware Asset
      operationId: createAsset
      tags:
        - Facilities Assets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HardwareAssetCreate'
      responses:
        '201':
          description: Hardware asset created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HardwareAsset'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
  /facilities/assets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get Hardware Asset Details
      operationId: getAssetById
      tags:
        - Facilities Assets
      responses:
        '200':
          description: Hardware asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HardwareAsset'
        '404':
          $ref: '#/components/responses/404NotFound'

    put:
      summary: Update Hardware Asset Details
      operationId: updateAsset
      tags:
        - Facilities Assets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HardwareAssetUpdate'
      responses:
        '200':
          description: Hardware asset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HardwareAsset'
        '404':
          $ref: '#/components/responses/404NotFound'

    delete:
      summary: Decommission Hardware Asset
      operationId: deleteAsset
      tags:
        - Facilities Assets
      responses:
        '204':
          description: Hardware asset decommissioned
...
  /facilities/sync/maintenance-logs:
    post:
      summary: Batch Sync Offline Maintenance Logs
      operationId: syncMaintenanceLogs
      tags:
        - Facilities Maintenance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/MaintenanceLogCreate'
      responses:
        '200':
          description: Offline maintenance logs synced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced_count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MaintenanceLog'
  /facilities/firmware:
    get:
      summary: List Firmware Releases
      operationId: listFirmwareReleases
      tags:
        - Facilities Firmware
      parameters:
        - name: device_type
          in: query
          required: true
          schema:
            type: string
            enum: [edge_camera, observation_perch, smart_collar, gateway, feeder]
      responses:
        '200':
          description: List of firmware releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FirmwareRelease'

    post:
      summary: Publish Firmware Release
      operationId: createFirmwareRelease
      tags:
        - Facilities Firmware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FirmwareReleaseCreate'
      responses:
        '201':
          description: Firmware release created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirmwareRelease'
  /facilities/firmware/ota-jobs:
    get:
      summary: List Device OTA Jobs
      operationId: listOTAJobs
      tags:
        - Facilities Firmware
      parameters:
        - name: asset_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of OTA jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceOTAJob'

    post:
      summary: Trigger OTA Firmware Update Pipeline
      operationId: triggerOTAUpdate
      tags:
        - Facilities Firmware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTATriggerRequest'
      responses:
        '202':
          description: OTA update pipeline triggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceOTAJob'
...
components:
  schemas:
    HardwareAsset:
      type: object
      required:
        - asset_id
        - serial_number
        - asset_type
        - model
        - status
        - created_at
      properties:
        asset_id:
          type: string
          format: uuid
        serial_number:
          type: string
        asset_type:
          type: string
          enum: [edge_camera, observation_perch, smart_collar, gateway, feeder]
        model:
          type: string
        status:
          type: string
          enum: [active, maintenance_required, offline, pending_return, decommissioned]
        zone_id:
          type: string
          format: uuid
          nullable: true
        assigned_employee_id:
          type: string
          format: uuid
          nullable: true
        last_ping_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    HardwareAssetCreate:
      type: object
      required:
        - serial_number
        - asset_type
        - model
      properties:
        serial_number:
          type: string
        asset_type:
          type: string
          enum: [edge_camera, observation_perch, smart_collar, gateway, feeder]
        model:
          type: string
        zone_id:
          type: string
          format: uuid
    HardwareAssetUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, maintenance_required, offline, pending_return, decommissioned]
        zone_id:
          type: string
          format: uuid
        assigned_employee_id:
          type: string
          format: uuid
...
    MaintenanceLog:
      type: object
      required:
        - log_id
        - asset_id
        - qr_code_scanned
        - action_taken
        - synced_at
        - created_at
      properties:
        log_id:
          type: string
          format: uuid
        ticket_id:
          type: string
          format: uuid
          nullable: true
        asset_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
          nullable: true
        qr_code_scanned:
          type: string
        action_taken:
          type: string
        notes:
          type: string
          nullable: true
        synced_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    MaintenanceLogCreate:
      type: object
      required:
        - asset_id
        - qr_code_scanned
        - action_taken
      properties:
        ticket_id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        qr_code_scanned:
          type: string
        action_taken:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
    FirmwareRelease:
      type: object
      required:
        - release_id
        - device_type
        - version
        - file_url
        - checksum
        - status
        - created_at
      properties:
        release_id:
          type: string
          format: uuid
        device_type:
          type: string
          enum: [edge_camera, observation_perch, smart_collar, gateway, feeder]
        version:
          type: string
        file_url:
          type: string
        checksum:
          type: string
        min_hardware_version:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, deprecated]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FirmwareReleaseCreate:
      type: object
      required:
        - device_type
        - version
        - file_url
        - checksum
      properties:
        device_type:
          type: string
          enum: [edge_camera, observation_perch, smart_collar, gateway, feeder]
        version:
          type: string
        file_url:
          type: string
        checksum:
          type: string
        min_hardware_version:
          type: string
        status:
          type: string
          enum: [draft, published, deprecated]
    DeviceOTAJob:
      type: object
      required:
        - job_id
        - asset_id
        - release_id
        - status
        - scheduled_at
        - created_at
      properties:
        job_id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        release_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, downloading, applying, completed, failed]
        error_message:
          type: string
          nullable: true
        scheduled_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OTATriggerRequest:
      type: object
      required:
        - asset_ids
        - release_id
      properties:
        asset_ids:
          type: array
          items:
            type: string
            format: uuid
        release_id:
          type: string
          format: uuid
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    '400BadRequest':
      description: Invalid request parameters or payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '401Unauthorized':
      description: Unauthenticated or invalid session cookie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '404NotFound':
      description: Requested entity not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    '500InternalError':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'